startserver.sh: Restore curl option to follow redirects
This option was added in #176 and then removed in #207 without any
explanation about what was wrong with it.

With curl 7.74.0 at least, the github URL in this script produces HTML
output and not a useable jar file unless you have this option set.

For example:
```
$ curl https://github.com/TeamAOF/ServerStarter/releases/download/v2.1.1/serverstarter-2.1.1.jar
<html><body>You are being <a href="https://objects.githubusercontent.com/github-production-release-asset-2e65be/357643340/5edf8551-5330-4e16-a986-601b858a10be?X-Amz-Algorithm=AWS4-HMAC-SHA256&amp;X-Amz-Credential=AKIAIWNJYAX4CSVEH53A%2F20220329%2Fus-east-1%2Fs3%2Faws4_request&amp;X-Amz-Date=20220329T050335Z&amp;X-Amz-Expires=300&amp;X-Amz-Signature=4828096047dea7ec1959ec7dfd835ba8fac2ac045e43efb92cdd29d4f0d391e9&amp;X-Amz-SignedHeaders=host&amp;actor_id=0&amp;key_id=0&amp;repo_id=357643340&amp;response-content-disposition=attachment%3B%20filename%3Dserverstarter-2.1.1.jar&amp;response-content-type=application%2Foctet-stream">redirected</a>.</body></html>
```

I think the issue fixed in #207 had nothing to do with this curl flag,
but the shebang that was also reverted in that commit.

Signed-off-by: Jim Ramsay <i.am@jimramsay.com>